TransferFunctionEditor, ColorMapPanel: Move all colormap stuff into one file out of TFEditor and put a nice border around it
